T3/requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
scikit-learn
matplotlib
seaborn
numpy
pandas